Add disengaged

add_disengaged() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add disengaged. The function returns Data (an output object of multiple potential types).

Usage

add_disengaged(
  data_xx,
  date_1L_chr,
  category_1L_chr = "Healthcare",
  date_tfmn_fn = lubridate::ymd,
  date_var_1L_chr = "Date",
  description_1L_chr = "Disengaged",
  dict_by_ctg_1L_chr = FALSE,
  uid_1L_chr = "UID",
  var_nm_1L_chr = "Disengaged"
)

Arguments

data_xx

Data (an output object of multiple potential types)

date_1L_chr

Date (a character vector of length one)

category_1L_chr

Category (a character vector of length one), Default: 'Healthcare'

date_tfmn_fn

Date transformation (a function), Default: lubridate::ymd

date_var_1L_chr

Date variable (a character vector of length one), Default: 'Date'

description_1L_chr

Description (a character vector of length one), Default: 'Disengaged'

dict_by_ctg_1L_chr

Dictionary by category (a character vector of length one), Default: FALSE

uid_1L_chr

Unique identifier (a character vector of length one), Default: 'UID'

var_nm_1L_chr

Variable name (a character vector of length one), Default: 'Disengaged'

Value

Data (an output object of multiple potential types)