Add episodes

add_episodes() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add episodes. The function returns Data (an output object of multiple potential types).

Usage

add_episodes(
  data_xx,
  separation_after_dbl,
  index_1L_int = integer(0),
  active_var_1L_chr = "Active",
  activity_var_1L_chr = "Activity",
  date_tfmn_fn = lubridate::ymd,
  date_var_1L_chr = "Date",
  end_date_dtm = NULL,
  episode_var_1L_chr = "Episodes",
  episodes_vars_chr = character(0),
  exclude_chr = "Duration",
  fiscal_start_1L_int = 7L,
  metrics_chr = make_metric_vars(),
  prefix_1L_chr = "Cumulative",
  separations_var_1L_chr = "Separations",
  temporal_vars_chr = make_temporal_vars(),
  uid_1L_chr = "UID",
  unit_1L_chr = "month"
)

Arguments

data_xx

Data (an output object of multiple potential types)

separation_after_dbl

Separation after (a double vector)

index_1L_int

Index (an integer vector of length one), Default: integer(0)

active_var_1L_chr

Active variable (a character vector of length one), Default: 'Active'

activity_var_1L_chr

Activity variable (a character vector of length one), Default: 'Activity'

date_tfmn_fn

Date transformation (a function), Default: lubridate::ymd

date_var_1L_chr

Date variable (a character vector of length one), Default: 'Date'

end_date_dtm

End date (a date vector), Default: NULL

episode_var_1L_chr

Episode variable (a character vector of length one), Default: 'Episodes'

episodes_vars_chr

Episodes variables (a character vector), Default: character(0)

exclude_chr

Exclude (a character vector), Default: 'Duration'

fiscal_start_1L_int

Fiscal start (an integer vector of length one), Default: 7

metrics_chr

Metrics (a character vector), Default: make_metric_vars()

prefix_1L_chr

Prefix (a character vector of length one), Default: 'Cumulative'

separations_var_1L_chr

Separations variable (a character vector of length one), Default: 'Separations'

temporal_vars_chr

Temporal variables (a character vector), Default: make_temporal_vars()

uid_1L_chr

Unique identifier (a character vector of length one), Default: 'UID'

unit_1L_chr

Unit (a character vector of length one), Default: 'month'

Value

Data (an output object of multiple potential types)