Add new unique identifier

add_new_uid() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add new unique identifier. The function returns Data (a tibble).

Usage

add_new_uid(
  data_tb,
  uid_vars_chr,
  uid_pfx_1L_chr,
  arrange_by_1L_chr = character(0),
  drop_old_uids_1L_lgl = FALSE,
  new_uid_var_1L_chr = "UID",
  imputed_uid_pfx_chr = "UNK",
  place_first_1L_lgl = TRUE,
  recode_1L_lgl = FALSE,
  recode_pfx_1L_chr = "Person_"
)

Arguments

data_tb

Data (a tibble)

uid_vars_chr

Unique identifier variables (a character vector)

uid_pfx_1L_chr

Unique identifier prefix (a character vector of length one)

arrange_by_1L_chr

Arrange by (a character vector of length one), Default: character(0)

drop_old_uids_1L_lgl

Drop old unique identifiers (a logical vector of length one), Default: FALSE

new_uid_var_1L_chr

New unique identifier variable (a character vector of length one), Default: 'UID'

imputed_uid_pfx_chr

Imputed unique identifier prefix (a character vector), Default: 'UNK'

place_first_1L_lgl

Place first (a logical vector of length one), Default: TRUE

recode_1L_lgl

Recode (a logical vector of length one), Default: FALSE

recode_pfx_1L_chr

Recode prefix (a character vector of length one), Default: 'Person_'

Value

Data (a tibble)