Add sampled variable

add_sampled_variable() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add sampled variable. The function returns Clients (a tibble).

Usage

add_sampled_variable(
  clients_tb,
  seed_ds_tb,
  shares_dbl,
  prefix_1L_chr = "Client_",
  scale_1L_dbl = 100,
  uid_var_nm_1L_chr = "Client ID",
  var_nm_1L_chr = "Sex"
)

Arguments

clients_tb

Clients (a tibble)

seed_ds_tb

Seed dataset (a tibble)

shares_dbl

Shares (a double vector)

prefix_1L_chr

Prefix (a character vector of length one), Default: 'Client_'

scale_1L_dbl

Scale (a double vector of length one), Default: 100

uid_var_nm_1L_chr

Unique identifier variable name (a character vector of length one), Default: 'Client ID'

var_nm_1L_chr

Variable name (a character vector of length one), Default: 'Sex'

Value

Clients (a tibble)