Get metric for range

get_metric_for_range() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get metric for range. The function returns Metric (a double vector of length one).

Usage

get_metric_for_range(
  series_tsb,
  metric_1L_chr,
  date_end_dtm = NULL,
  date_start_dtm = NULL,
  frequency_1L_chr = c("daily", "weekly", "monthly", "quarterly", "yearly", "fiscal"),
  summary_fn = sum
)

Arguments

series_tsb

Series (a tsibble)

metric_1L_chr

Metric (a character vector of length one)

date_end_dtm

Date end (a date vector), Default: NULL

date_start_dtm

Date start (a date vector), Default: NULL

frequency_1L_chr

Frequency (a character vector of length one), Default: c("daily", "weekly", "monthly", "quarterly", "yearly", "fiscal")

summary_fn

Summary (a function), Default: sum

Value

Metric (a double vector of length one)