Get providers by date

get_providers_by_date() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get providers by date. The function returns Summary (a list).

Usage

get_providers_by_date(
  providers_tb,
  by_1L_chr = "year",
  clients_tb = NULL,
  counts_1L_lgl = FALSE,
  date_vars_chr = c("Retainer date", "Date"),
  index_1L_lgl = TRUE,
  index_day_1L_chr = "01",
  index_month_1L_chr = "07",
  end_dtm = NULL,
  provider_id_1L_chr = "ProviderID",
  start_dtm = NULL,
  type_1L_chr = c("all", "new", "cumulative"),
  what_1L_chr = c("activity", "census", "composite")
)

Arguments

providers_tb

Providers (a tibble)

by_1L_chr

By (a character vector of length one), Default: 'year'

clients_tb

Clients (a tibble), Default: NULL

counts_1L_lgl

Counts (a logical vector of length one), Default: FALSE

date_vars_chr

Date variables (a character vector), Default: c("Retainer date", "Date")

index_1L_lgl

Index (a logical vector of length one), Default: TRUE

index_day_1L_chr

Index day (a character vector of length one), Default: '01'

index_month_1L_chr

Index month (a character vector of length one), Default: '07'

end_dtm

End (a date vector), Default: NULL

provider_id_1L_chr

Provider identity (a character vector of length one), Default: 'ProviderID'

start_dtm

Start (a date vector), Default: NULL

type_1L_chr

Type (a character vector of length one), Default: c("all", "new", "cumulative")

what_1L_chr

What (a character vector of length one), Default: c("activity", "census", "composite")

Value

Summary (a list)