Get raw Estimatedesident Population data

get_raw_erp_data() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get raw estimatedesident population data. The function returns Estimatedesident Population raw (a tibble).

Usage

get_raw_erp_data(
  uid_1L_chr = "ERP_Q",
  age_chr = as.character(1:115),
  end_1L_chr = character(0),
  frequency_1L_chr = "quarterly",
  measure_chr = c("count", "change", "%change"),
  region_chr = c("NSW", "VIC", "QLD", "SA", "WA", "TAS", "NT", "ACT", "OT", "AUS"),
  sex_chr = c("female", "male", "person"),
  start_1L_chr = character(0),
  version_1L_chr = NULL
)

Arguments

uid_1L_chr

Unique identifier (a character vector of length one), Default: 'ERP_Q'

age_chr

Age (a character vector), Default: as.character(1:115)

end_1L_chr

End (a character vector of length one), Default: character(0)

frequency_1L_chr

Frequency (a character vector of length one), Default: 'quarterly'

measure_chr

Measure (a character vector), Default: c("count", "change", "%change")

region_chr

Region (a character vector), Default: c("NSW", "VIC", "QLD", "SA", "WA", "TAS", "NT", "ACT", "OT", "AUS")

sex_chr

Sex (a character vector), Default: c("female", "male", "person")

start_1L_chr

Start (a character vector of length one), Default: character(0)

version_1L_chr

Version (a character vector of length one), Default: NULL

Value

Estimatedesident Population raw (a tibble)