Get temporal function

get_temporal_fn() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get temporal function. The function returns Temporal (a function).

Usage

get_temporal_fn(
  period_1L_chr = make_temporal_vars(index_1L_chr = "Sub"),
  temporal_fns_ls = NULL,
  daily_fn = make_date_tfmn_fn(),
  monthly_fn = tsibble::yearmonth,
  fiscal_start_1L_int = 7L
)

Arguments

period_1L_chr

Period (a character vector of length one), Default: make_temporal_vars(index_1L_chr = "Sub")

temporal_fns_ls

Temporal functions (a list), Default: NULL

daily_fn

Daily (a function), Default: make_date_tfmn_fn()

monthly_fn

Monthly (a function), Default: tsibble::yearmonth

fiscal_start_1L_int

Fiscal start (an integer vector of length one), Default: 7

Value

Temporal (a function)