Get variables

get_vars() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get variables. The function returns Variables (a character vector).

Usage

get_vars(
  data_df,
  activity_1L_chr = "Activity",
  appointments_var_1L_chr = "Appointments",
  cancellations_var_1L_chr = "Cancellations",
  clinical_team_1L_chr = "Clinical Team",
  clinician_1L_chr = "Clinician",
  clinician_discipline_1L_chr = "Service",
  components_chr = c("Year", "Quarter", "Week"),
  cost_var_1L_chr = "Cost",
  days_1L_chr = "Weekday",
  duration_1L_chr = "Duration",
  exclude_chr = character(0),
  group_1L_chr = character(0),
  index_1L_chr = "Date",
  referrals_var_1L_chr = "Referrals",
  referrers_1L_chr = "Referrer Role",
  severity_1L_chr = "Severity",
  team_disciplines_1L_chr = "Disciplines",
  uid_var_1L_chr = "UID",
  what_1L_chr = c("all", "clinical", "metrics", "sports", "temporal", "other")
)

Arguments

data_df

Data (a data.frame)

activity_1L_chr

Activity (a character vector of length one), Default: 'Activity'

appointments_var_1L_chr

Appointments variable (a character vector of length one), Default: 'Appointments'

cancellations_var_1L_chr

Cancellations variable (a character vector of length one), Default: 'Cancellations'

clinical_team_1L_chr

Clinical team (a character vector of length one), Default: 'Clinical Team'

clinician_1L_chr

Clinician (a character vector of length one), Default: 'Clinician'

clinician_discipline_1L_chr

Clinician discipline (a character vector of length one), Default: 'Service'

components_chr

Components (a character vector), Default: c("Year", "Quarter", "Week")

cost_var_1L_chr

Cost variable (a character vector of length one), Default: 'Cost'

days_1L_chr

Days (a character vector of length one), Default: 'Weekday'

duration_1L_chr

Duration (a character vector of length one), Default: 'Duration'

exclude_chr

Exclude (a character vector), Default: character(0)

group_1L_chr

Group (a character vector of length one), Default: character(0)

index_1L_chr

Index (a character vector of length one), Default: 'Date'

referrals_var_1L_chr

Referrals variable (a character vector of length one), Default: 'Referrals'

referrers_1L_chr

Referrers (a character vector of length one), Default: 'Referrer Role'

severity_1L_chr

Severity (a character vector of length one), Default: 'Severity'

team_disciplines_1L_chr

Team disciplines (a character vector of length one), Default: 'Disciplines'

uid_var_1L_chr

Unique identifier variable (a character vector of length one), Default: 'UID'

what_1L_chr

What (a character vector of length one), Default: c("all", "clinical", "metrics", "sports", "temporal", "other")

Value

Variables (a character vector)