Make Estimatedesident Population dataset

make_erp_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make estimatedesident population dataset. The function returns Estimatedesident Population (an output object of multiple potential types).

Usage

make_erp_ds(
  erp_raw_tb = get_raw_erp_data(region_chr = "AUS"),
  age_range_int = 1:115,
  age_bands_lup = NULL,
  age_tfmn_fn = as.integer,
  as_dyad_1L_lgl = TRUE,
  frequency_1L_chr = "quarterly",
  measure_1L_chr = "count",
  select_chr = c("time_period", "age", "sex", "obs_value"),
  sex_chr = c("Female", "Male", "Total"),
  summarise_1L_lgl = FALSE,
  var_ctg_chr = c("Temporal", "Key", "Key", "Metric")
)

Arguments

erp_raw_tb

Estimatedesident Population raw (a tibble), Default: get_raw_erp_data(region_chr = "AUS")

age_range_int

Age range (an integer vector), Default: 1:115

age_bands_lup

Age bands (a lookup table), Default: NULL

age_tfmn_fn

Age transformation (a function), Default: as.integer

as_dyad_1L_lgl

As dyad (a logical vector of length one), Default: TRUE

frequency_1L_chr

Frequency (a character vector of length one), Default: 'quarterly'

measure_1L_chr

Measure (a character vector of length one), Default: 'count'

select_chr

Select (a character vector), Default: c("time_period", "age", "sex", "obs_value")

sex_chr

Sex (a character vector), Default: c("Female", "Male", "Total")

summarise_1L_lgl

Summarise (a logical vector of length one), Default: FALSE

var_ctg_chr

Variable category (a character vector), Default: c("Temporal", "Key", "Key", "Metric")

Value

Estimatedesident Population (an output object of multiple potential types)