Make forecasts tibble

make_forecasts_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make forecasts tibble. The function returns Forecast (a tibble).

Usage

make_forecasts_tb(
  fabels_ls,
  group_by_1L_chr = character(0),
  group_fn = sum,
  summary_fn = mean,
  tfmn_args_ls = NULL,
  tfmn_fn = NULL,
  tfmn_pattern_1L_chr = "Transformed_{.col}",
  type_1L_chr = c("default", "grouped", "summary", "both"),
  what_1L_chr = "Appointments"
)

Arguments

fabels_ls

Fabels (a list)

group_by_1L_chr

Group by (a character vector of length one), Default: character(0)

group_fn

Group (a function), Default: sum

summary_fn

Summary (a function), Default: mean

tfmn_args_ls

Transformation arguments (a list), Default: NULL

tfmn_fn

Transformation (a function), Default: NULL

tfmn_pattern_1L_chr

Transformation pattern (a character vector of length one), Default: 'Transformed_.col'

type_1L_chr

Type (a character vector of length one), Default: c("default", "grouped", "summary", "both")

what_1L_chr

What (a character vector of length one), Default: 'Appointments'

Value

Forecast (a tibble)