Make sampling lookup table

make_sampling_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make sampling lookup table. The function returns Sampling (a lookup table).

Usage

make_sampling_lup(
  shares_dbl,
  values_xx,
  var_nm_1L_chr,
  prefix_1L_chr = "Client_",
  scale_1L_dbl = 100,
  uid_var_nm_1L_chr = "Client ID"
)

Arguments

shares_dbl

Shares (a double vector)

values_xx

Values (an output object of multiple potential types)

var_nm_1L_chr

Variable name (a character vector of length one)

prefix_1L_chr

Prefix (a character vector of length one), Default: 'Client_'

scale_1L_dbl

Scale (a double vector of length one), Default: 100

uid_var_nm_1L_chr

Unique identifier variable name (a character vector of length one), Default: 'Client ID'

Value

Sampling (a lookup table)