Make scaled forecasts

make_scaled_forecasts() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make scaled forecasts. The function returns Forecasts (a tibble).

Usage

make_scaled_forecasts(
  forecasts_tb,
  predictors_tb,
  after_1L_chr = character(0),
  before_1L_chr = character(0),
  bind_1L_lgl = TRUE,
  positive_1L_lgl = FALSE,
  prefix_1L_chr = "scenario_",
  reference_1L_chr = "Status quo",
  scale_fn = function(x) x * 1,
  scaled_fn_ls = NULL,
  tfmn_1_fn = identity,
  tfmn_2_fn = identity
)

Arguments

forecasts_tb

Forecasts (a tibble)

predictors_tb

Predictors (a tibble)

after_1L_chr

After (a character vector of length one), Default: character(0)

before_1L_chr

Before (a character vector of length one), Default: character(0)

bind_1L_lgl

Bind (a logical vector of length one), Default: TRUE

positive_1L_lgl

Positive (a logical vector of length one), Default: FALSE

prefix_1L_chr

Prefix (a character vector of length one), Default: 'scenario_'

reference_1L_chr

Reference (a character vector of length one), Default: 'Status quo'

scale_fn

Scale (a function), Default: function(x) x * 1

scaled_fn_ls

Scaled (a list of functions), Default: NULL

tfmn_1_fn

Transformation 1 (a function), Default: identity

tfmn_2_fn

Transformation 2 (a function), Default: identity

Value

Forecasts (a tibble)