Make scenario forecast costs

make_scenario_forecast_costs() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make scenario forecast costs. The function returns Forecast costs (a tibble).

Usage

make_scenario_forecast_costs(
  scenario_forecasts_ls,
  unit_costs_tb,
  what_1L_chr,
  fixed_cost_1L_dbl = 0,
  positive_1L_lgl = FALSE,
  predictors_chr = character(0),
  type_1L_chr = c("Variable", "Fixed")
)

Arguments

scenario_forecasts_ls

Scenario forecasts (a list)

unit_costs_tb

Unit costs (a tibble)

what_1L_chr

What (a character vector of length one)

fixed_cost_1L_dbl

Fixed cost (a double vector of length one), Default: 0

positive_1L_lgl

Positive (a logical vector of length one), Default: FALSE

predictors_chr

Predictors (a character vector), Default: character(0)

type_1L_chr

Type (a character vector of length one), Default: c("Variable", "Fixed")

Value

Forecast costs (a tibble)