Make summary dataset

make_summary_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make summary dataset. The function returns Data (an output object of multiple potential types).

Usage

make_summary_ds(
  data_xx,
  active_base_1L_chr = "Active",
  add_with_join_xx = NULL,
  join_before_dtm = NULL,
  index_1L_chr = "Date",
  key_vars_chr = character(0),
  max_tenure_1L_dbl = numeric(0),
  metrics_chr = make_metric_vars(),
  patterns_ls = NULL,
  prefix_1L_chr = "Cumulative",
  separation_after_dbl = numeric(0),
  tenure_var_1L_chr = "Tenure",
  uid_1L_chr = "UID",
  update_desc_1L_lgl = TRUE
)

Arguments

data_xx

Data (an output object of multiple potential types)

active_base_1L_chr

Active base (a character vector of length one), Default: 'Active'

add_with_join_xx

Add with join (an output object of multiple potential types), Default: NULL

join_before_dtm

Join before (a date vector), Default: NULL

index_1L_chr

Index (a character vector of length one), Default: 'Date'

key_vars_chr

Key variables (a character vector), Default: character(0)

max_tenure_1L_dbl

Maximum tenure (a double vector of length one), Default: numeric(0)

metrics_chr

Metrics (a character vector), Default: make_metric_vars()

patterns_ls

Patterns (a list), Default: NULL

prefix_1L_chr

Prefix (a character vector of length one), Default: 'Cumulative'

separation_after_dbl

Separation after (a double vector), Default: numeric(0)

tenure_var_1L_chr

Tenure variable (a character vector of length one), Default: 'Tenure'

uid_1L_chr

Unique identifier (a character vector of length one), Default: 'UID'

update_desc_1L_lgl

Update description (a logical vector of length one), Default: TRUE

Value

Data (an output object of multiple potential types)