Make time series models list

make_ts_models_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make time series models list. The function returns Time series models (a list).

Usage

make_ts_models_ls(
  mabels_ls = list(),
  args_ls = make_tfmn_args_ls(),
  cumulatives_args_ls = make_tfmn_args_ls(),
  fabels_ls = list(),
  join_to_args_ls = make_tfmn_args_ls(),
  predictor_args_ls = make_tfmn_args_ls(),
  models_chr = character(0),
  test_1L_int = integer(0)
)

Arguments

mabels_ls

Mabels (a list), Default: list()

args_ls

Arguments (a list), Default: make_tfmn_args_ls()

cumulatives_args_ls

Cumulatives arguments (a list), Default: make_tfmn_args_ls()

fabels_ls

Fabels (a list), Default: list()

join_to_args_ls

Join to arguments (a list), Default: make_tfmn_args_ls()

predictor_args_ls

Predictor arguments (a list), Default: make_tfmn_args_ls()

models_chr

Models (a character vector), Default: character(0)

test_1L_int

Test (an integer vector of length one), Default: integer(0)

Value

Time series models (a list)