Plot forecast

plot_forecast() is a Plot function that plots data. Specifically, this function implements an algorithm to plot forecast. The function is called for its side effects and does not return a value.

Usage

plot_forecast(
  data_xx,
  ts_models_ls,
  metrics_chr = make_metric_vars(),
  x_label_1L_chr = ""
)

Arguments

data_xx

Data (an output object of multiple potential types)

ts_models_ls

Time series models (a list)

metrics_chr

Metrics (a character vector), Default: make_metric_vars()

x_label_1L_chr

X label (a character vector of length one), Default: ”

Value

No return value, called for side effects.