Transform to tsibble

transform_to_tsibble() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform to tsibble. The function returns Data (a tsibble).

Usage

transform_to_tsibble(
  data_tb,
  activity_1L_chr = "Activity",
  athlete_roles_chr = c("Athlete", "AlumniAthlete"),
  appointments_var_1L_chr = "Appointments",
  cancellations_var_1L_chr = "Cancellations",
  clinical_team_1L_chr = "Clinical Team",
  clinician_1L_chr = "Clinician",
  clinician_discipline_1L_chr = "Service",
  components_chr = c("Year", "Quarter", "Week"),
  cost_var_1L_chr = "Cost",
  date_tfmn_fn = identity,
  days_1L_chr = "Weekday",
  duration_1L_chr = "Duration",
  exclude_chr = "Group",
  fiscal_start_1L_int = 7L,
  group_1L_chr = character(0),
  index_1L_chr = "Date",
  is_wide_1L_lgl = F,
  key_vars_chr = character(0),
  metrics_chr = make_metric_vars(),
  referrals_var_1L_chr = "Referrals",
  referrers_1L_chr = "Referrer Role",
  severity_1L_chr = "Severity",
  team_disciplines_1L_chr = "Disciplines",
  temporal_vars_chr = make_temporal_vars(),
  uid_var_1L_chr = "UID",
  type_1L_chr = c("main", "focused"),
  what_1L_chr = c("all", "totals")
)

Arguments

data_tb

Data (a tibble)

activity_1L_chr

Activity (a character vector of length one), Default: 'Activity'

athlete_roles_chr

Athlete roles (a character vector), Default: c("Athlete", "AlumniAthlete")

appointments_var_1L_chr

Appointments variable (a character vector of length one), Default: 'Appointments'

cancellations_var_1L_chr

Cancellations variable (a character vector of length one), Default: 'Cancellations'

clinical_team_1L_chr

Clinical team (a character vector of length one), Default: 'Clinical Team'

clinician_1L_chr

Clinician (a character vector of length one), Default: 'Clinician'

clinician_discipline_1L_chr

Clinician discipline (a character vector of length one), Default: 'Service'

components_chr

Components (a character vector), Default: c("Year", "Quarter", "Week")

cost_var_1L_chr

Cost variable (a character vector of length one), Default: 'Cost'

date_tfmn_fn

Date transformation (a function), Default: identity

days_1L_chr

Days (a character vector of length one), Default: 'Weekday'

duration_1L_chr

Duration (a character vector of length one), Default: 'Duration'

exclude_chr

Exclude (a character vector), Default: 'Group'

fiscal_start_1L_int

Fiscal start (an integer vector of length one), Default: 7

group_1L_chr

Group (a character vector of length one), Default: character(0)

index_1L_chr

Index (a character vector of length one), Default: 'Date'

is_wide_1L_lgl

Is wide (a logical vector of length one), Default: F

key_vars_chr

Key variables (a character vector), Default: character(0)

metrics_chr

Metrics (a character vector), Default: make_metric_vars()

referrals_var_1L_chr

Referrals variable (a character vector of length one), Default: 'Referrals'

referrers_1L_chr

Referrers (a character vector of length one), Default: 'Referrer Role'

severity_1L_chr

Severity (a character vector of length one), Default: 'Severity'

team_disciplines_1L_chr

Team disciplines (a character vector of length one), Default: 'Disciplines'

temporal_vars_chr

Temporal variables (a character vector), Default: make_temporal_vars()

uid_var_1L_chr

Unique identifier variable (a character vector of length one), Default: 'UID'

type_1L_chr

Type (a character vector of length one), Default: c("main", "focused")

what_1L_chr

What (a character vector of length one), Default: c("all", "totals")

Value

Data (a tsibble)